requirements.txt
numpy
scipy
matplotlib
pandas
sklearn
nltk
notebook
nbimporter
keras
tensorflow
progressbar2